refactor: index파일 내에서 export
