+++ src/apis/hooks/useGetFlower.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import { flower } from '@/apis/service/flower.ts'
 import { FlowerInfo } from '@/types'
 
-export const useGetFlower = ({ id }: { id: number }) => {
+const useGetFlower = ({ id }: { id: number }) => {
   const queryKey = ['flower']
   const { isLoading, isError, data } = useQuery<FlowerInfo, Error>({
     queryKey,
@@ -11,3 +11,5 @@ export const useGetFlower = ({ id }: { id: number }) => {
 
   return { isLoading, isError, data }
 }
+
+export default useGetFlower
