+++ src/apis/hooks/useGetLetterSent.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import { letterSent } from '@/apis/service/letterSent.ts'
 import { LetterSentInfo } from '@/types'
 
-export const useGetLetterSent = () => {
+const useGetLetterSent = () => {
   const queryKey = ['letterSent']
   const { isLoading, isError, data } = useQuery<LetterSentInfo, Error>({
     queryKey,
@@ -11,3 +11,5 @@ export const useGetLetterSent = () => {
 
   return { isLoading, isError, data }
 }
+
+export default useGetLetterSent
