+++ src/apis/service/letterSent.ts
@@ -1,9 +1,11 @@
 import apiClient from '@/apis/apiClient.ts'
 
-export const letterSent = async () => {
+const letterSent = async () => {
   try {
     return await apiClient.get('/letter-sent').then((res) => res.data)
   } catch (error) {
     console.log(error)
   }
 }
+
+export default letterSent
